Create R_File_Collection
This folder is for all of the R documentation, plotting, ideas, scripts etc. along with personal folders
All personal folders should be marked with *last_name_first_name*, i.e. *Palencia_Hansel*
This will allow for good organization, once we have a rough idea and understanding of what we are wanting to do we can move it to the production folder.